+++ backend/plugin/src/main/java/app/epistola/valtimo/config/EpistolaPluginFactory.java
@@ -2,6 +2,7 @@ package app.epistola.valtimo.config;
 
 import app.epistola.valtimo.mapping.JsonataMappingService;
 import app.epistola.valtimo.service.EpistolaService;
+import app.epistola.valtimo.service.completion.EpistolaResultCollectorRunner;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ritense.document.service.DocumentService;
 import com.ritense.plugin.PluginFactory;
@@ -15,25 +16,28 @@ public class EpistolaPluginFactory extends PluginFactory<EpistolaPlugin> {
     private final ObjectMapper objectMapper;
     private final JsonataMappingService jsonataMappingService;
     private final DocumentService documentService;
+    private final EpistolaResultCollectorRunner resultCollectorRunner;
 
     public EpistolaPluginFactory(
             @NotNull PluginService pluginService,
             @NotNull EpistolaService epistolaService,
             @NotNull ObjectMapper objectMapper,
             @NotNull JsonataMappingService jsonataMappingService,
-            @NotNull DocumentService documentService
+            @NotNull DocumentService documentService,
+            @NotNull EpistolaResultCollectorRunner resultCollectorRunner
     ) {
         super(pluginService);
         this.epistolaService = epistolaService;
         this.objectMapper = objectMapper;
         this.jsonataMappingService = jsonataMappingService;
         this.documentService = documentService;
+        this.resultCollectorRunner = resultCollectorRunner;
     }
 
     @NotNull
     @Override
     protected EpistolaPlugin create() {
         return new EpistolaPlugin(epistolaService, objectMapper,
-                jsonataMappingService, documentService);
+                jsonataMappingService, documentService, resultCollectorRunner);
     }
 }